% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adheaping-package.R
\docType{package}
\name{adheaping-package}
\alias{adheaping}
\alias{adheaping-package}
\title{adheaping: characteristic-function de-heaping density estimation}
\description{
Tuning-free kernel density estimation for heaped and rounded data. See the vignette
\code{vignette("adheaping")}. The methods derive from the spectral-decomposition kernel
density estimation of Thornton (arXiv:2606.15450). The main entry point is \code{\link{adkde}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mitch-thornton/kde-ad-heaping}
  \item Report bugs at \url{https://github.com/mitch-thornton/kde-ad-heaping/issues}
}

}
\author{
\strong{Maintainer}: Mitchell A. Thornton \email{mitch@smu.edu}

}
\keyword{internal}
